+++ backend/Domain/Models/Nutrients.cs
@@ -5,15 +5,15 @@ namespace Domain.Models;
 
 public class Nutrients : ITableDataBase
 {
-    public int? Id { get; set; }
-    [JsonPropertyName("menu_item_id")] public int? MenuItemId { get; set; }
+    public long? Id { get; set; }
+    [JsonPropertyName("menu_item_id")] public long? MenuItemId { get; set; }
     public int? Kilocalories { get; set; }
     public int? Proteins { get; set; }
     public int? Fats { get; set; }
     public int? Carbohydrates { get; set; }
 
     public Nutrients(){}
-    public Nutrients(int? id, int? kilocalories, int? proteins, int? fats, int? carbohydrates)
+    public Nutrients(long? id, int? kilocalories, int? proteins, int? fats, int? carbohydrates)
     {
         Id = id;
         Kilocalories = kilocalories;
